+++ src/main/java/com/example/spot/auth/application/refactor/impl/OAuthMemberProcessor.java
@@ -1,8 +1,12 @@
-package com.example.spot.auth.application.refactor.member;
+package com.example.spot.auth.application.refactor.impl;
 
 import com.example.spot.auth.application.refactor.TokenProvider;
 import com.example.spot.auth.application.refactor.dto.OAuthProfile;
 import com.example.spot.auth.application.refactor.dto.SocialAccountResult;
+import com.example.spot.auth.application.refactor.member.OAuthMemberConflictProcessor;
+import com.example.spot.auth.application.refactor.member.OAuthMemberCreator;
+import com.example.spot.auth.application.refactor.member.ProfileCompletenessChecker;
+import com.example.spot.auth.application.refactor.member.RefreshTokenStore;
 import com.example.spot.auth.presentation.dto.token.TokenResponseDTO.TokenDTO;
 import com.example.spot.member.domain.Member;
 import com.example.spot.member.presentation.dto.MemberResponseDTO;
@@ -17,7 +21,7 @@ public class OAuthMemberProcessor {
     private final OAuthMemberCreator oAuthMemberCreator;
     private final OAuthMemberConflictProcessor oAuthMemberConflictProcessor;
     private final ProfileCompletenessChecker profileCompletenessChecker;
-    private final TokenProvider tokenProvider; // TODO 인터페이스 분리
+    private final TokenProvider tokenProvider;
     private final RefreshTokenStore refreshTokenStore;
 
     @Transactional
